+++ docs/frontend.md
@@ -1,0 +1,3 @@
+## Ragnarok Frontend
+
+We use Gradio to provide the UI
